xapp-gtk-window: Trap X errors when setting progress
This should be a best effort feature for applications
and it shouldn't ever result in making applications
segfault.

Note: This could be tested easily by setting progress on
an erroneous XID. This resulted in a BadWindow crash.
With this commit, the error is now trapped and no impact
is suffered on the calling application.