xapp-icon-chooser-button.c: make dialog inherit modal state (#64)
This makes the XAppIconChooserDialog inherit the modal state from
the toplevel widget of the icon chooser button. This fixes an issue
where the dialog is unresponsive when the button is pressed while
in a modal window - such as the dialog that opens when adding a new
entry to the editable list widget in Cinnamon's xlet settings.