status icons: Add a monitor class for applets - XAppStatusIconMonitor, (#70)
and encapsulate all dbus work inside the two main xapp
classes.

The monitor is used by an applet (that displays icons) to receive
added and removed signals, when an XAppStatusIcon joins or leaves
the bus.

The interface used by status icons (org.x.StatusIcon) is now generated
and the XAppStatusIcon class uses its skeleton (server) interface,
while the monitor creates a proxy for each icon that appears on the
bus, sending this to the applet.  This proxy can be treated like
any gobject, and represents the status icon to the applet.

Updated test scripts to cover new functionality, fixed some of the
doc generation (more to do still).