mate applet: fix configuring
We need to replace @libexec@, so first use configure_file() and then
feed the result to i18n.

While we are at it, also rename the files to end in .in as relevant.
This is a useful flag to indicate they need to be configured, and/or
how many times they need to be configured. def_file doesn't actually
change via i18n anyway, but ¯\_(ツ)_/¯ I assume it will eventually get
them.