xapp-status-icon: Track and notify changes in types of support for the icon - notify when a status icon is picked up by a native applet, a traditional system tray, or is not being picked up by anything.
FIXME: Cinnamon currently doesn't notify when the tray applet is removed because
the underlying plug/socket are still retained in cinnamon's tray subsystem.  We
should maybe eventually propagate the applet no longer displaying icons down to
cinnamon-tray-manager and na-tray-manager to remove the icon entirely (this is
done correctly in mate-panel and xfce4).