xapp-status-icon.c: Reposition the menu when part of it would end up below the workarea of the monitor.
The "push_in" option in the Gtk's menu positioning function only
moves the outer box of the menu within limits, not its contents.

In single monitor configurations, this fix will move the entire menu
while accounting for panels.  In multi-monitor situations, it may
fall back to only keeping it inside the monitor bounds (this is a
limitation in Gdk and the EWMH spec).