xapp-status-icon.c: Only call gtk_status_icon_set_name() when explicitly requested.
This destroys and re-creates the status icon plug, causing Cinnamon's tray
applet to think it's a new icon.  Also, the name is write-only, so we can't
check whether it's actually changed or not.

The other GtkStatusIcon calls (icon, visible, label, tooltip) don't have
this effect, but we'll still check on our end if they've changed before
passing it on to Gtk.

Fixes linuxmint/blueberry#99