Add StatusNotifier watcher.
See: https://www.freedesktop.org/wiki/Specifications/StatusNotifierItem/

This is a backgroud program that is started when an XAppStatusIconMonitor
is created (by cinnamon's or mate's xapp applet for instance).

It advertises as a StatusNotifierWatcher and allows programs that implement
StatusNotifierIcon to register, and an XAppStatusIcon will be created and
sent to the applet.  This includes appindicator-based icons as well.

Closes linuxmint/cinnamon#9176