Rework dbus naming conventions to be more friendly with sandboxed environments.
- XAppStatusIcon users will own the name 'org.x.StatusIcon.progname".
  This can also be specified using xapp_status_icon_set_name(), to
  assign a particular name, as long as there are 4 parts to the name
  in total.
- Each app process will have a GDBusObjectManagerServer singleton
  to expose individual icon objects (this keeps the ability to have
  multiple status icons from a single process, something xapp-sn-
  watcher must have). The server is created when the first icon is
  added, and destroyed when the last is removed.
- XAppStatusIconMonitors will host a matching object manager client
  for each status icon process it connects with.

This simplifies accounting of icons, as the managers take over
add/remove notifications and proxy creation - we really only need
to keep track of the managers themselves and react to them.

Minor changes to status applets are required to ensure the visual
representations are tracked correctly. Previously, each icon had
a matching unique name on the bus. This name was used as the
identifier to keep track of proxies. Now, there is one name per
process, so the object path will be appended also, to insure
uniqueness.