xapp-sn-watcher.c: Don't try to use a null pointer as a hash table lookup key.
When a new item tries to register, its key is added to the item table
with a null value, to prevent trying to register the same client
multiple times. When registration is finished, the key is removed,
and re-used to add the new proxy back into the table.

If the item disappears before that registration finished, we remove
it from the table (handle_sn_item_name_owner_lost), so when the item
registration completes (it will because a proxy doesn't need an owner),
our steal/re-add fails because the key was removed already.

This will have us steal the key early when completing registration,
and abort if it's null.

Also wrap g_object_unref when removing an item from the table to
prevent a warning.