xapp-gtk3-module.c: Apply window icon override to all windows for an app.
Firefox 92 creates and destroys multiple windows at startup, but
this module was only overriding the icon for the initial one.

This has a potential impact for additional windows opened by
the user deliberately (right-click an external link and open
in new window) - these windows will end up with the same icon.

Fixes #145