Revert "xapp-sn-watcher: Don't enable in MATE"
This reverts commit b6b856496d4c2de66dcb05ce4c5225c0fee118c4.

Tested in 24.04.

The following applets do not support SNI:

- mate-indicator-applet
- mate-indicator-applet-complete
- mate-panel/notification-area-applet

mate-xapp-status-applet only supports it if xapp-sn-watcher is running,
and it won't if MATE isn't in this list.

Note: Ayatana indicators do support SNI but they're not used in MATE
by default. mate-xapp-status-applet is recommended instead. If people
want to use Ayatana indicators, they need to remove MATE from this list
so the two do not conflict.