Revert "Remove XAppKbdLayoutController class."
This reverts commit c22f968cb324528c95edce03cc1c7f3b2e176e83.

Even though it's gone from Cinnamon and nobody uses it, we should keep
this for another year or so.

- This gives us flexibility in tagging. We don't want xapp to have
to wait for Cinnamon to be ready to be tagged (This is  important for
rolling distros).

- This gives us flexibility in backporting. New tools might depend on
xapp 3.0. We want to be able to backport them with the lib, without
having to backport Cinnamon.